Update schema: add username to profiles, change set_weight to float
- Add unique username column to profiles table
- Change set_weight from int to float to support fractional weights
- Add corresponding Supabase migrations for both changes
- Add CLAUDE.md and globals.d.ts to .gitignore

+++ supabase/schemas/workout_sets.sql
@@ -4,7 +4,7 @@ create table public.workout_sets (
     exercise_id int references exercises,
     set_number int,
     reps int not null default 0,
-    set_weight int not null default 0,
+    set_weight float not null default 0,
     time_seconds int, --seconds
     distance float, --miles
     notes text,
